Move to sending DEVICE-ID as a header
Make the API closer to what it'll look like when we have authentication.

Require the device id to be sent when interacting with ratings such that
each rating is "owned" by a device and can only be deleted/updated by
the same device and the index only returns ratings by the given device.

Make it optional when fetching resources (if it is set return the rating
for the given device for the given resource).

Remove unused resource routes until we need them.