Introduce Algolia incremental index job
This job scans through all of the resources that have been updated since
the last time it ran. Approved resources and their approved services are
updated in Algolia, and all others are removed from Algolia.

We introduce a new table named `bookmarks` that allows the job to pick
up again from where it left off the last time it ran.

Updated ChangeRequestsController to make it so the `updated_at` time of
the associated Resource is touched whenever a ChangeRequest is applied.
Allows the Algolia incremental index job to detect updates.

Remove the deprecated `Delayed::Job` system. It has been replaced for a
while now by the k8s CronJob setup.